improved progress bar rounding and added AbortExecute
