+++ pyvista/utilities/__init__.py
@@ -1,6 +1,6 @@
 """Utilities routines."""
 
-from .errors import (GPUInfo, Observer, ProgressMonitor, Report,
+from .errors import (GPUInfo, Observer, Report,
                      assert_empty_kwargs, get_gpu_info, send_errors_to_logging,
                      set_error_output_file)
 from .features import *
