+++ tests/test_utilities.py
@@ -201,3 +201,12 @@ def test_convert_id_list():
         id_list.SetId(i, v)
     converted = helpers.vtk_id_list_to_array(id_list)
     assert np.allclose(converted, ids)
+
+
+def test_progress_monitor():
+    # dragon = ex.download_dragon()
+    mesh = ex.download_bunny()
+    ugrid = mesh.delaunay_3d(progress_bar=True)
+
+
+test_progress_monitor()
